- move to proper test class testCreateUserWithOnlyRequiredAttributesAndVerifiesStatusCode201
