+++ scimono-compliance-tests/src/main/java/com/sap/scimono/scim/system/tests/E2EUserComplianceTest.java
@@ -30,7 +30,6 @@ import static com.sap.scimono.scim.system.tests.util.TestUtil.constructResourceL
 import static java.util.Collections.singletonMap;
 import static javax.servlet.http.HttpServletResponse.SC_NOT_FOUND;
 import static javax.servlet.http.HttpServletResponse.SC_NO_CONTENT;
-import static javax.ws.rs.core.Response.Status.CREATED;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -1037,18 +1036,6 @@ public class E2EUserComplianceTest extends SCIMComplianceTest {
             ((EnterpriseExtension) fetchedEmployeeUser.getExtension(ENTERPRISE_URN)).getManager().getValue(), "Verify manager 'value'")
     );
     // @formatter:on
-  }
-
-  @Test
-  @DisplayName("Test create a user with only required attributes and verifies, that the status code is 201")
-  public void testCreateUserWithOnlyRequiredAttributesAndVerifiesStatusCode201(){
-    String testUserName = "testCreateUserWithOnlyRequiredAttributesAndVerifiesStatusCode201";
-    User testUser = TestData.setAttributesToATestUser(testUserName).build();
-
-    logger.info("Creating User: {}, with username", testUserName);
-    SCIMResponse<User> scimResponse = resourceAwareUserRequest.createUser(testUser);
-
-    assertAll("Verify Create User Response", getResponseStatusAssertions(scimResponse, true, CREATED));
   }
 
   private Collection<Executable> getFullUserAssertions(final User expected, final User actual) {
